refactor dircopy to separate util
