+++ solr/test-framework/src/java/org/apache/solr/EmbeddedSolrServerTestBase.java
@@ -28,15 +28,11 @@ import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.Properties;
 
-import org.apache.commons.io.FileUtils;
-import org.apache.lucene.util.LuceneTestCase;
 import org.apache.solr.client.solrj.embedded.EmbeddedSolrServer;
 import org.apache.solr.common.util.ContentStream;
 import org.apache.solr.common.util.ContentStreamBase;
 import org.apache.solr.common.util.ContentStreamBase.ByteArrayStream;
-import org.apache.solr.util.ExternalPaths;
 import org.junit.After;
 import org.junit.AfterClass;
 
@@ -121,32 +117,9 @@ abstract public class EmbeddedSolrServerTestBase extends SolrTestCaseJ4 {
   }
 
   public static void initCore() throws Exception {
-    final String home = legacyExampleCollection1SolrHome();
+    final String home = SolrJettyTestBase.legacyExampleCollection1SolrHome();
     final String config = home + "/" + DEFAULT_CORE_NAME + "/conf/solrconfig.xml";
     final String schema = home + "/" + DEFAULT_CORE_NAME + "/conf/schema.xml";
     initCore(config, schema, home);
   }
-
-  public static String legacyExampleCollection1SolrHome() throws IOException {
-    final String sourceHome = ExternalPaths.SOURCE_HOME;
-    if (sourceHome == null)
-      throw new IllegalStateException("No source home! Cannot create the legacy example solr home directory.");
-
-    final Path tempSolrHome = LuceneTestCase.createTempDir();
-    Files.copy(Path.of(sourceHome, "server", "solr", "solr.xml"), tempSolrHome.resolve("solr.xml"));
-    final Path collectionDir = tempSolrHome.resolve(DEFAULT_CORE_NAME);
-    Files.createDirectories(collectionDir);
-    final File configSetDir = new File(sourceHome, "server/solr/configsets/sample_techproducts_configs/conf");
-    FileUtils.copyDirectoryToDirectory(configSetDir, collectionDir.toFile());
-
-    final Properties props = new Properties();
-    props.setProperty("name", DEFAULT_CORE_NAME);
-
-    try (OutputStream outputStream = Files.newOutputStream(collectionDir.resolve(CORE_PROPERTIES_FILENAME))) {
-      props.store(outputStream, null);
-    }
-
-    return tempSolrHome.toAbsolutePath().toString();
-  }
-
 }
